pyproject.toml
[tool.black]
line-length = 128
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 128
skip = [".venv"]

[tool.semantic_release]
version_variable = ["setup.py:version"]
branch = "main"
changelog_file = "CHANGELOG.md"
build_command = "python -m build"
dist_path = "dist/"
upload_to_repository = false
commit_version_number = true
tag_commit = true
push_to_remote = false
commit_message = "chore(release): {version}\n\nAutomatically generated by python-semantic-release"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:bravado_core.spec",
    "ignore::DeprecationWarning:swagger_spec_validator",
    "ignore::DeprecationWarning:bravado_core.swagger20_validator"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecosystems-cli"
version = "1.0.0"
description = "CLI for ecosyste.ms API"
authors = [
    {name = "Sebastian Schürmann", email = "sebastian.schurmann@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click",
    "requests",
    "pyyaml",
    "rich",
    "mcp",
    "bravado",
    "bravado-core",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "black", "isort", "flake8", "bandit", "complexipy"]

[project.scripts]
ecosystems = "ecosystems_cli.cli:main"

[project.urls]
Homepage = "https://github.com/ecosyste-ms/ecosyste_ms_cli"
Repository = "https://github.com/ecosyste-ms/ecosyste_ms_cli"
